Renamed app class and test and package name
